@@ -43,70 +43,72 @@
   },
   "homepage": "https://github.com/richardcyrus/fm-rmx-product-feedback-app#readme",
   "dependencies": {
-    "@prisma/client": "^5.8.0",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@remix-run/css-bundle": "2.5.1",
-    "@remix-run/node": "2.5.1",
-    "@remix-run/react": "2.5.1",
-    "@remix-run/serve": "2.5.1",
-    "@vercel/analytics": "^1.1.2",
-    "@vercel/remix": "^2.5.1",
-    "isbot": "^4.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tiny-invariant": "^1.3.1",
-    "zod": "^3.22.4"
+    "@prisma/client": "^5.22.0",
+    "@radix-ui/react-label": "^2.1.8",
+    "@radix-ui/react-select": "^2.2.6",
+    "@radix-ui/react-tabs": "^1.1.13",
+    "@react-router/fs-routes": "^7.11.0",
+    "@react-router/node": "^7.11.0",
+    "@react-router/serve": "^7.11.0",
+    "isbot": "^5.1.32",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
+    "react-router": "^7.11.0",
+    "tiny-invariant": "^1.3.3",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
-    "@axe-core/playwright": "^4.8.3",
-    "@faker-js/faker": "^8.3.1",
-    "@playwright/test": "^1.41.1",
-    "@remix-run/dev": "2.5.1",
+    "@axe-core/playwright": "^4.11.0",
+    "@eslint/compat": "^2.0.0",
+    "@eslint/js": "^9.39.2",
+    "@faker-js/faker": "^10.1.0",
+    "@playwright/test": "^1.57.0",
+    "@react-router/dev": "^7.11.0",
     "@svgr/cli": "^8.1.0",
     "@svgr/plugin-jsx": "^8.1.0",
     "@svgr/plugin-prettier": "^8.1.0",
     "@svgr/plugin-svgo": "^8.1.0",
-    "@testing-library/jest-dom": "^6.2.0",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/eslint": "^8.56.2",
-    "@types/node": "^20.11.0",
-    "@types/react": "^18.2.47",
-    "@types/react-dom": "^18.2.18",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.16",
-    "axe-core": "^4.8.3",
-    "concurrently": "^8.2.2",
-    "cssnano": "^6.0.3",
-    "dotenv-cli": "^7.3.0",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "happy-dom": "^13.0.6",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.1",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/node": "^22",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
+    "@vitejs/plugin-react": "^5.1.2",
+    "autoprefixer": "^10.4.23",
+    "axe-core": "^4.11.0",
+    "concurrently": "^9.2.1",
+    "cssnano": "^7.1.2",
+    "dotenv-cli": "^11.0.0",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import-x": "^4.16.1",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-react-refresh": "^0.4.26",
+    "globals": "^16.5.0",
+    "happy-dom": "^20.0.11",
     "husky": "^8.0.3",
-    "lint-staged": "^15.2.0",
-    "micromatch": "^4.0.5",
-    "postcss": "^8.4.33",
-    "postcss-cli": "^11.0.0",
-    "postcss-import": "^16.0.0",
-    "postcss-nested": "^6.0.1",
+    "lint-staged": "^15.5.2",
+    "micromatch": "^4.0.8",
+    "postcss": "^8.5.6",
+    "postcss-cli": "^11.0.1",
+    "postcss-import": "^16.1.1",
+    "postcss-nested": "^7.0.2",
     "postcss-sort-media-queries": "^5.2.0",
-    "prettier": "^3.2.2",
-    "prisma": "^5.8.0",
-    "stylelint": "^16.1.0",
-    "stylelint-config-standard": "^36.0.0",
-    "stylelint-order": "^6.0.4",
-    "tsx": "^4.7.0",
-    "typescript": "^5.1.6",
-    "vite-tsconfig-paths": "^4.2.3",
-    "vitest": "^1.2.0"
+    "prettier": "^3.7.4",
+    "prisma": "^5.22.0",
+    "stylelint": "^16.26.1",
+    "stylelint-config-standard": "^39.0.1",
+    "stylelint-order": "^7.0.1",
+    "tsx": "^4.21.0",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.51.0",
+    "vite": "^7.3.0",
+    "vite-tsconfig-paths": "^6.0.3",
+    "vitest": "^4.0.16"
   },
   "browserslist": {
     "production": [
@@ -122,10 +124,11 @@
     ]
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=24.12.0"
   },
   "sideEffects": false,
   "volta": {
     "node": "24.12.0"
-  }
+  },
+  "packageManager": "pnpm@10.26.2+sha512.0e308ff2005fc7410366f154f625f6631ab2b16b1d2e70238444dd6ae9d630a8482d92a451144debc492416896ed16f7b114a86ec68b8404b2443869e68ffda6"
 }
