@@ -1,6 +1,6 @@
-const micromatch = require("micromatch");
+import micromatch from "micromatch";
 
-module.exports = {
+export default {
   "*": (allFiles) => {
     const commands = [];
 
@@ -14,30 +14,28 @@ module.exports = {
     const codeFiles = micromatch(allFiles, ["**/*.[tj]s?(x)"]);
     if (codeFiles.length > 0) {
       commands.push(
-        `eslint --cache --cache-location ./node_modules/.cache/eslint --fix --ignore-path .gitignore ${codeFiles.join(
-          " "
-        )}`
+        `eslint --fix ${codeFiles.join(" ")}`
       );
     }
 
-    // Lint and fix PostCSS files with Stylelint
-    const postCssFiles = micromatch(allFiles, ["**/*.pcss"]);
+    // Format Markdown, JSON, and CSS files.
+    const miscFiles = micromatch(allFiles, [
+      "**/*.md",
+      "**/*.json",
+      "**/*.css",
+    ]);
+    if (miscFiles.length > 0) {
+      commands.push(`prettier --write ${miscFiles.join(" ")}`);
+    }
+
+    // Lint and fix CSS files with Stylelint
+    const postCssFiles = micromatch(allFiles, ["**/*.css"]);
     if (postCssFiles.length > 0) {
       commands.push(
         `stylelint ${postCssFiles.join(
           " "
         )} --cache --cache-location ./node_modules/.cache/stylelint --fix`
       );
-    }
-
-    // Format Markdown, JSON, and PostCSS files.
-    const miscFiles = micromatch(allFiles, [
-      "**/*.md",
-      "**/*.json",
-      "**/*.pcss",
-    ]);
-    if (miscFiles.length > 0) {
-      commands.push(`prettier --write ${miscFiles.join(" ")}`);
     }
 
     return commands;
