@@ -1,5 +1,5 @@
-import AxeBuilder from "@axe-core/playwright";
-import { test, expect } from "@playwright/test";
+import { AxeBuilder } from "@axe-core/playwright";
+import { expect, test } from "@playwright/test";
 
 test.describe("Test accessibility of the Roadmap page", () => {
   test("should not have any detectable accessibility issues", async ({
